move user affiliation loader tests
